move conjugate_mul to ops.hpp
