+++ include/graphblas/algorithms/norm.hpp
@@ -41,65 +41,6 @@
 
 namespace grb {
 
-	namespace internal {
-
-		template<
-			typename IN1, typename IN2, typename OUT,
-			enum Backend implementation = config::default_backend
-		>
-		class conjugate_mul {
-
-			public:
-
-				typedef IN1 left_type;
-				typedef IN2 right_type;
-				typedef OUT result_type;
-
-				static constexpr bool has_foldl = false;
-				static constexpr bool has_foldr = false;
-				static constexpr bool is_associative = false;
-				static constexpr bool is_commutative = false;
-
-				/**
-				 * Out-of-place application of the multiplication c = a * b.
-				 *
-				 * @param[in]  a Pointer to the left-hand side input. Must be initialised.
-				 * @param[in]  b Pointer to the right-hand side input. Must be initialised.
-				 * @param[out] c Pointer to where to compute the output.
-				 *
-				 * \warning All pointers must be valid or UB occurs.
-				 */
-				static void apply(
-						const IN1 * __restrict__ const a,
-						const IN2 * __restrict__ const b,
-						OUT * __restrict__ const c
-				) {
-					*c = *a * grb::utils::is_complex< IN2 >::conjugate( *b );
-				}
-		};
-	} // namespace internal
-
-	template<
-		typename IN1, typename IN2, typename OUT,
-		enum Backend implementation = config::default_backend
-	>
-	class conjugate_mul : public operators::internal::Operator<
-		internal::conjugate_mul< IN1, IN2, OUT, implementation >
-	> {
-
-	public:
-
-		template< typename A, typename B, typename C, enum Backend D >
-		using GenericOperator = internal::conjugate_mul< A, B, C, D >;
-		conjugate_mul() {}
-	};
-
-
-	template< typename D1, typename D2, typename D3, enum Backend implementation >
-	struct is_operator< conjugate_mul< D1, D2, D3, implementation > > {
-		static const constexpr bool value = true;
-	};
-
 	namespace algorithms {
 
 		template< typename OutputType, typename InputType >
@@ -144,7 +85,7 @@ namespace grb {
 
 			RC ret = grb::dot< descr >(
 				yyt, y, y, ring.getAdditiveMonoid(),
-				conjugate_mul< InputType, InputType, InputType >()
+				grb::operators::conjugate_mul< InputType, InputType, InputType >()
 			);
 			if( ret == SUCCESS ) {
 				Semiring<
